REFACTOR: controller structure
