+++ src/main/java/segelzwerg/sporttooolbox/web/pace/PaceService.java
@@ -1,24 +1,26 @@
-package segelzwerg.sporttooolbox.Controller;
+package segelzwerg.sporttooolbox.web.pace;
 
 import org.springframework.stereotype.Component;
 import segelzwerg.sporttooolbox.IUnits.Distance;
-import segelzwerg.sporttooolbox.IUnits.Speed;
+import segelzwerg.sporttooolbox.IUnits.Pace;
 import segelzwerg.sporttooolbox.IUnits.Time;
 import segelzwerg.sporttooolbox.SpeedCalculator;
+import segelzwerg.sporttooolbox.web.speed.SpeedForm;
+
 
 /**
- * Speed calculating service
+ * Pace calculating service
  */
 @Component
-public class SpeedService {
+public class PaceService {
 
     /**
-     * Calculate speed
+     * calulates pace
      *
-     * @param form form with values
-     * @return calculated speed
+     * @param form
+     * @return calculated pace
      */
-    public Speed calculateSpeed(SpeedForm form) {
+    public Pace calculatePace(SpeedForm form) {
         String majorUnit = form.getDistanceMajorUnit();
         String minorUnit = form.getDistanceMinorUnit();
         int major = form.getMajor();
@@ -32,8 +34,6 @@ public class SpeedService {
 
         SpeedCalculator speedCalculator = new SpeedCalculator(distance, time);
 
-        return speedCalculator.computeSpeed();
+        return speedCalculator.computePace();
     }
-
-
 }
