+++ src/main/java/segelzwerg/sporttooolbox/web/speed/SpeedController.java
@@ -1,4 +1,4 @@
-package segelzwerg.sporttooolbox.Controller;
+package segelzwerg.sporttooolbox.web.speed;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -6,6 +6,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import segelzwerg.sporttooolbox.IUnits.Speed;
+import segelzwerg.sporttooolbox.web.Translator;
 
 /**
  * Speed Controller
