+++ src/main/java/segelzwerg/sporttooolbox/web/speed/SpeedForm.java
@@ -1,4 +1,4 @@
-package segelzwerg.sporttooolbox.Controller;
+package segelzwerg.sporttooolbox.web.speed;
 
 import lombok.Data;
